backend/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
PyJWT==2.3.0